pyproject.toml
# #############################################################################
[tool.poetry]
name        = "m3"
version     = "0.1.0"
description = ""
authors     = ["Rafael <rafael.cerezo.martin@icloud.com>"]
license     = "MIT"
readme      = "README.md"
packages    = [{include = "src"}]

# #############################################################################
# PRODUCTION DEPENDENCIES
# poetry install --without dev
# #############################################################################

[tool.poetry.dependencies]
python = ">=3.12,<3.14"

# Manipulation et analyse de données scientifiques
pandas          = "^2.3.3"          # Manipulation de DataFrames et analyse de données
numpy           = "2.2.2"           # Calcul numérique et tableaux multidimensionnels
scipy           = "^1.16.3"         # Algorithmes scientifiques et mathématiques avancés

# Machine Learning et statistiques
scikit-learn    = "^1.7.2"          # Bibliothèque de machine learning classique
statsmodels     = "^0.14.5"         # Modèles statistiques et tests d'hypothèses

# Visualisation de données
matplotlib      = "<=3.10.0"         # Création de graphiques et visualisations
seaborn         = "^0.13.2"         # Visualisations statistiques de haut niveau

# Configuration et environnement
python-dotenv   = "^1.2.1"          # Gestion des variables d'environnement (.env)

# Deep Learning et NLP (nécessite Windows)
transformers    = { version = "^4.57.3", markers = "sys_platform == 'win32'" } # Modèles de transformers (BERT, GPT, etc.)
huggingface-hub = { version = "<1.0.0",  markers = "sys_platform == 'win32'" } # Interface avec Hugging Face Hub
datasets        = { version = "^4.4.1",  markers = "sys_platform == 'win32'" } # Chargement et traitement de datasets

# Vision par ordinateur (nécessite Windows)
opencv-python-headless = { version = "^4.12.0.88", markers = "sys_platform == 'win32'" }  # Traitement d'images sans GUI

# Utilitaires
tqdm            = "^4.67.1"         # Barres de progression pour boucles


# #############################################################################
# DEVELOPPEMENT DEPENDENCIES
# poetry install --with dev
# #############################################################################
ydata-profiling = {extras = ["notebook"], version = "^4.18.0"}
jupyter-contrib-nbextensions = "^0.7.0"
ipywidgets = "^8.1.8"
polars = "^1.36.1"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
# Environnement Jupyter pour notebooks interactifs
jupyter         = "^1.1.1"          # Suite complète Jupyter
jupyterlab      = "^4.5.0"          # Interface web moderne pour notebooks
ipykernel       = "^7.1.0"          # Noyau Python pour Jupyter

# Qualité de code et formatage
pre-commit      = "^4.5.0"          # Hooks Git pour vérifications automatiques
black           = "^25.11.0"        # Formatage automatique de code Python
flake8          = "^7.3.0"          # Vérification de style et erreurs

# Utilitaires de développement
requests        = "^2.32.5"         # Requêtes HTTP simplifiées
pillow          = "^12.0.0"         # Traitement d'images PIL

# Deep Learning (dev - nécessite Windows)
transformers    = { version = "^4.57.3", markers = "sys_platform == 'win32'" } # Modèles de transformers pour expérimentation

# #############################################################################
[build-system]
requires      = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"